aPort rules to Gsuite Activity Event
