+++ gsuite_activityevent_rules/gsuite_mobile_device_compromise.yml
@@ -1,0 +1,74 @@
+AnalysisType: rule
+Filename: gsuite_mobile_device_compromise.py
+RuleID: GSuite.DeviceCompromise
+DisplayName: GSuite User Device Compromised
+Enabled: true
+LogTypes:
+  - GSuite.ActivityEvent
+Tags:
+  - GSuite
+Severity: Medium
+Description: >
+  GSuite reported a user's device has been compromised.
+Reference: https://developers.google.com/admin-sdk/reports/v1/appendix/activity/mobile#DEVICE_COMPROMISED_EVENT
+Runbook: >
+  Have the user change their passwords and reset the device.
+SummaryAttributes:
+  - actor
+Tests:
+  -
+    Name: Normal Mobile Event
+    ExpectedResult: false
+    Log:
+      {
+        "id": {
+          "applicationName": "mobile",
+        },
+        "actor": {
+          "callerType": "USER",
+          "email": "homer.simpson@example.io",
+        },
+        "type": "device_updates",
+        "name": "DEVICE_REGISTER_UNREGISTER_EVENT",
+        "parameters": {
+          "USER_EMAIL": "homer.simpson@example.io"
+        },
+      }
+  -
+    Name: Suspicious Activity Shows not Compromised
+    ExpectedResult: false
+    Log:
+      {
+        "id": {
+          "applicationName": "mobile",
+        },
+        "actor": {
+          "callerType": "USER",
+          "email": "homer.simpson@example.io",
+        },
+        "type": "device_updates",
+        "name": "DEVICE_COMPROMISED_EVENT",
+        "parameters": {
+          "USER_EMAIL": "homer.simpson@example.io",
+          "DEVICE_COMPROMISED_STATE": "NOT_COMPROMISED"
+        },
+      }
+  -
+    Name: Suspicious Activity Shows Compromised
+    ExpectedResult: true
+    Log:
+      {
+        "id": {
+          "applicationName": "mobile",
+        },
+        "actor": {
+          "callerType": "USER",
+          "email": "homer.simpson@example.io",
+        },
+        "type": "device_updates",
+        "name": "DEVICE_COMPROMISED_EVENT",
+        "parameters": {
+          "USER_EMAIL": "homer.simpson@example.io",
+          "DEVICE_COMPROMISED_STATE": "COMPROMISED"
+        },
+      }
