+++ gsuite_activityevent_rules/gsuite_mobile_device_suspicious_activity.py
@@ -1,12 +1,11 @@
 from panther_base_helpers import deep_get
-from panther_base_helpers import gsuite_details_lookup as details_lookup
 
 
 def rule(event):
     if deep_get(event, "id", "applicationName") != "mobile":
         return False
 
-    return bool(details_lookup("suspicious_activity", ["SUSPICIOUS_ACTIVITY_EVENT"], event))
+    return bool(event.get("name") == "SUSPICIOUS_ACTIVITY_EVENT")
 
 
 def title(event):
