--- gsuite_reports_rules/gsuite_mobile_device_compromise.yml
@@ -1,57 +1,0 @@
-AnalysisType: rule
-Filename: gsuite_mobile_device_compromise.py
-RuleID: GSuite.DeviceCompromise
-DisplayName: GSuite User Device Compromised
-Enabled: true
-LogTypes:
-  - GSuite.Reports
-Tags:
-  - GSuite
-Severity: Medium
-Description: >
-  GSuite reported a user's device has been compromised.
-Reference: https://developers.google.com/admin-sdk/reports/v1/appendix/activity/mobile#DEVICE_COMPROMISED_EVENT
-Runbook: >
-  Have the user change their passwords and reset the device.
-SummaryAttributes:
-  - actor
-Tests:
-  -
-    Name: Normal Mobile Event
-    ExpectedResult: false
-    Log:
-      {
-        'actor': {'email': 'bobert@example.com'},
-        'id': {'applicationName': 'mobile'},
-        'events': [{'type': 'device_updates'}]
-      }
-  -
-    Name: Suspicious Activity Shows not Compromised
-    ExpectedResult: false
-    Log:
-      {
-        'actor': {'email': 'bobert@example.com'},
-        'id': {'applicationName': 'mobile'},
-        'events': [
-          {
-            'type': 'suspicious_activity',
-            'name': 'DEVICE_COMPROMISED_EVENT',
-            'parameters': [{'name': 'DEVICE_COMPROMISED_STATE', 'value': 'NOT_COMPROMISED'}]
-          }
-        ]
-      }
-  -
-    Name: Suspicious Activity Shows Compromised
-    ExpectedResult: true
-    Log:
-      {
-        'actor': {'email': 'bobert@example.com'},
-        'id': {'applicationName': 'mobile'},
-        'events': [
-          {
-            'type': 'suspicious_activity',
-            'name': 'DEVICE_COMPROMISED_EVENT',
-            'parameters': [{'name': 'DEVICE_COMPROMISED_STATE', 'value': 'COMPROMISED'}]
-          }
-        ]
-      }
